Added all-in-one query endpoint.

+++ src/storage/secondary_index.rs
@@ -8,31 +8,42 @@ pub trait SecondaryIndex<PrimaryStoreSchema>
         PrimaryStoreSchema: KeyValueSchema<Key=<Self as KeyValueSchema>::Value>,
 {
     type FieldType;
-    fn accessor(value: &PrimaryStoreSchema::Value) -> Self::FieldType;
+    fn accessor(value: &PrimaryStoreSchema::Value) -> Option<Self::FieldType>;
     fn make_index(key: &PrimaryStoreSchema::Key, value: Self::FieldType) -> Self::Key;
     fn make_prefix_index(value: Self::FieldType) -> Self::Key;
 
     fn store_index(&self, key: &PrimaryStoreSchema::Key, value: &PrimaryStoreSchema::Value) -> Result<(), StorageError> {
-        let index = Self::make_index(key, Self::accessor(value));
         let db = self.as_ref();
-        Ok(db.put(&index, key)?)
+        if let Some(field) = Self::accessor(value) {
+            let index = Self::make_index(key, field);
+            Ok(db.put(&index, key)?)
+        } else {
+            Ok(())
+        }
     }
 
     fn delete_index(&self, key: &PrimaryStoreSchema::Key, value: &PrimaryStoreSchema::Value) -> Result<(), StorageError> {
-        let index = Self::make_index(key, Self::accessor(value));
         let db = self.as_ref();
-        Ok(db.delete(&index)?)
+        if let Some(field) = Self::accessor(value) {
+            let index = Self::make_index(key, field);
+            Ok(db.delete(&index)?)
+        } else {
+            Ok(())
+        }
     }
 
     fn get_index(&self, key: &PrimaryStoreSchema::Key, value: &PrimaryStoreSchema::Value) -> Result<Option<PrimaryStoreSchema::Key>, StorageError> {
-        let index = Self::make_index(key, Self::accessor(value));
         let db = self.as_ref();
-        Ok(db.get(&index)?)
+        if let Some(field) = Self::accessor(value) {
+            let index = Self::make_index(key, field);
+            Ok(db.get(&index)?)
+        } else {
+            Ok(None)
+        }
     }
 
     fn get_raw_prefix_iterator(&self, field: Self::FieldType) -> Result<IteratorWithSchema<Self>, StorageError> {
         let prefix = Self::make_prefix_index(field);
-
         Ok(self.as_ref().prefix_iterator(&prefix)?)
     }
 }
